Refactor docker and oci format parsing into their own files
